fixed final issues in plots

+++ comments/jfactor_plots/alternate_scaled_plot/calc.py
@@ -2,10 +2,14 @@
 from math import log10, exp, pi, atan, cos, sqrt
 from scipy.integrate   import quad
 from scipy.interpolate import interp1d
-import matplotlib.pyplot as plt
-import veripy
+import matplotlib.pyplot  as plt
+import matplotlib.patches as patches
+import veripy, aplpy, os, sys
 from astropy.io import fits
+from matplotlib import rc
+rc('text',usetex=True)
 
+fast    = True
 output  = 'output/'
 r_s     = 15.14 # kpc
 alpha   =  0.17
@@ -14,6 +18,7 @@ rho_sol =  0.4  # GeV/cm^3
 rad2deg = 180 / pi
 deg2rad = pi / 180
 kpc2cm  = 3.086e21
+force   = ( '-f' in sys.argv )
 def unnorm_einasto(r) :
   f = r / r_s
   f = f**alpha - 1
@@ -23,28 +28,30 @@ rho_s = rho_sol / unnorm_einasto( r_sol )
 def einasto(r) : 
   return rho_s * unnorm_einasto(r)
 
-rmin   = 1e-3 # kpc
-rmax   = 1    # kpc
-nradii = 100
-rdelta = ( log10(rmax) - log10(rmin) ) / (nradii-1)
-radii  = [ pow(10, log10(rmin) + i * rdelta ) for i in range(nradii) ]
-angles = [ atan(r/r_sol) * rad2deg for r in radii ]
-dens   = [ einasto(r) for r in radii ]
 fname  = output + 'calc_densities.pdf'
-fig    = plt.figure()
-fax    = fig.add_subplot(111)
-#plt.plot( radii, dens, label='einasto' )
-#plt.xlabel( 'Radius from Galactic Center [kpc]' )
-plt.plot( angles, dens, label='einasto' )
-plt.xlabel( 'Angle from Galactic Center [deg]' )
-plt.ylabel( 'Mass Density [GeV/cm^3]' )
-plt.title( 'Dark Matter Mass Density around Galactic Center' )
-plt.xscale('log')
-plt.yscale('log')
-legend = fax.legend( loc='lower left', shadow=True )
-fax.grid()
-plt.savefig( fname, dpi=300 )
-plt.clf()
+if not os.path.exists( fname ) :
+  rmin   = 1e-3 # kpc
+  rmax   = 1    # kpc
+  nradii = 100
+  rdelta = ( log10(rmax) - log10(rmin) ) / (nradii-1)
+  radii  = [ pow(10, log10(rmin) + i * rdelta ) for i in range(nradii) ]
+  angles = [ atan(r/r_sol) * rad2deg for r in radii ]
+  dens   = [ einasto(r) for r in radii ]
+  fig    = plt.figure()
+  fax    = fig.add_subplot(111)
+  #plt.plot( radii, dens, label='einasto' )
+  #plt.xlabel( 'Radius from Galactic Center [kpc]' )
+  plt.plot( angles, dens, label='einasto' )
+  
+  plt.xlabel( 'Angle from the Galactic Center [deg]' )
+  plt.ylabel( 'Mass Density [GeV/cm^3]' )
+  plt.title( 'Dark Matter Mass Density around the Galactic Center' )
+  plt.xscale('log')
+  plt.yscale('log')
+  legend = fax.legend( loc='lower left', shadow=True )
+  fax.grid()
+  plt.savefig( fname, dpi=300 )
+  plt.clf()
 
 def jfac(ang) :
   # ang : radians, angle from GC
@@ -66,19 +73,21 @@ angles  = [ pow(10, log10(amin) + i * adelta ) for i in range(nangles) ] # deg
 radians = [ a * deg2rad for a in angles ] # radians
 jfacs   = [ jfac(rad) for rad in radians ]
 fname   = output + 'calc_jfacs.pdf'
-fig     = plt.figure()
-fax     = fig.add_subplot(111)
-plt.plot( angles, jfacs, label='jfac' )
-plt.xlim( amin, amax )
-plt.xlabel( 'Angle from Galactic Center [deg]' )
-plt.ylabel( 'jfac [GeV^2/cm^5]' )
-plt.title( 'Dark Matter Mass Density around Galactic Center' )
-plt.xscale('log')
-plt.yscale('log')
-legend = fax.legend( loc='lower left', shadow=True )
-fax.grid()
-plt.savefig( fname, dpi=300 )
-plt.clf()
+if not os.path.exists( fname ) :
+  fig     = plt.figure()
+  fax     = fig.add_subplot(111)
+  plt.plot( angles, jfacs, label='jfac' )
+  
+  plt.xlim( amin, amax )
+  plt.xlabel( 'Angle from the Galactic Center [deg]' )
+  plt.ylabel( 'jfac [GeV^2/cm^5]' )
+  plt.title( 'Dark Matter Mass Density around the Galactic Center' )
+  plt.xscale('log')
+  plt.yscale('log')
+  legend = fax.legend( loc='lower left', shadow=True )
+  fax.grid()
+  plt.savefig( fname, dpi=300 )
+  plt.clf()
 
 def Jfactor(ang, wid) :
   # ang : radians, angle from GC
@@ -94,7 +103,7 @@ print('Integral: %g GeV^2/cm^5' % integ)
 
 a_min  = 1e-5
 a_max  = 5
-nangs  = 200
+nangs  = 1000
 adelta = ( log10(a_max) - log10(a_min) ) / (nangs-1)
 angles_interp  = [ pow(10, log10(a_min) + i * adelta ) for i in range(nangs) ]
 radians_interp = [ a*deg2rad for a in angles_interp ]
@@ -119,10 +128,11 @@ jfs_l_deg2 = [ jl * (pi**2) / (180**2)  for jl in jfs_l ]
 print('FITS J factor %g %s' % ( fits_jfactor, fits_jf_unit ) )
 interpf = interp1d( radians_interp, jfs_l )
 kern    = lambda rad : rad * interpf( rad )
-integ   = 2 * pi * quad( kern, 0*deg2rad, 3*deg2rad )[0]
-print( '\ninteg3: %g\n' % integ )
+if not fast :
+  integ   = 2 * pi * quad( kern, 0*deg2rad, 3*deg2rad )[0]
+  print( '\ninteg3: %g\n' % integ )
+print('post integral')
 
-fname = output + 'calc_Jfactor_per_deg2.pdf'
 wid = 1e-8 # radians
 #wid = 0.01 * 0.5 * deg2rad  # radians, 0.01 deg integration radius
 Jfactors_per_sr   = [ Jfactor(rad,wid) / Jfactor_area(rad,wid) for rad in radians ]
@@ -131,24 +141,46 @@ Jfactors_per_deg2 = [ j * (pi**2) / (180**2) for j in Jfactors_per_sr ]
 Jfactors_per_sr_integ = [ Jfactor(rad,wid) / Jfactor_area(rad,wid) for rad in radians_interp ]
 interpf = interp1d( radians_interp, Jfactors_per_sr_integ )
 kern    = lambda rad : rad * interpf( rad )
-integ   = 2 * pi * quad( kern, 0*deg2rad, 3*deg2rad )[0]
-print('theory jfactor %g GeV^2/cm^5' % integ )
+if not fast :
+  integ   = 2 * pi * quad( kern, 0*deg2rad, 3*deg2rad )[0]
+  print('theory jfactor %g GeV^2/cm^5' % integ )
+print('after jfactor calc')
 
+fname = output + 'calc_Jfactor_per_deg2.pdf'
+if force or not os.path.exists( fname ) :
+  fig     = plt.figure()
+  fax     = fig.add_subplot(111)
+  #plt.plot( angles, Jfactors_per_sr, label='Jfactors_per_sr' )
+  #plt.ylabel( 'J Factor per Solid Angle [GeV^2/cm^5/sr]' )
+  #plt.plot( angles, Jfactors_per_deg2, label='theory' )
+  plt.plot( angles_interp, jfs_l_deg2, label='template' )
+  #plt.ylabel( r'J Factor per Solid Angle $\left [\frac{GeV^2}{cm^5 deg^2} \right ]$' )
+  plt.ylabel( r'J-Factor per Solid Angle $\left [\mathrm{GeV}^{2} \; \mathrm{cm}^{-5} \; \mathrm{deg}^{-2} \right ]$' )
+  
+  # blue rectangular analysis region
+  b_xmin = amin
+  b_xwid = 3 # deg
+  b_ymin = 4e20 # GeV^2 cm^-5 deg^-2
+  b_ywid = ( max(jfs_l_deg2) * 1.5 ) - b_ymin
+  args  = [ (b_xmin, b_ymin) ]
+  args += [ b_xwid ]
+  args += [ b_ywid ]
+  kwa  = {}
+  kwa['linewidth'] = 0
+  kwa['facecolor'] = 'lightblue'
+  kwa['zorder'   ] = 1
+  kwa['label'    ] = 'Observation Radius'
+  fax.add_patch( patches.Rectangle( *args, **kwa ) )
+  
+  plt.xlabel( r'Angle from the Galactic Center $\left [ \mathrm{deg} \right ]$' )
+  plt.title( 'Dark Matter J-Factor' )
+  plt.xscale('log')
+  plt.yscale('log')
+  plt.xlim( 1e-2, 5 )
+  plt.ylim( b_ymin, b_ymin+b_ywid )
+  #legend = fax.legend( loc='lower left', shadow=True )
+  fax.grid()
+  plt.savefig( fname, bbox_inches='tight', dpi=300 )
 
-fig     = plt.figure()
-fax     = fig.add_subplot(111)
-#plt.plot( angles, Jfactors_per_sr, label='Jfactors_per_sr' )
-#plt.ylabel( 'J Factor per Solid Angle [GeV^2/cm^5/sr]' )
-plt.plot( angles, Jfactors_per_deg2, label='theory' )
-plt.plot( angles_interp, jfs_l_deg2, label='template' )
-plt.ylabel( r'J Factor per Solid Angle $\left [\frac{GeV^2}{cm^5 deg^2} \right ]$' )
-plt.xlabel( 'Angle from Galactic Center [deg]' )
-plt.title( 'Dark Matter J Factor per Solid Angle  around Galactic Center' )
-plt.xscale('log')
-plt.yscale('log')
-plt.xlim( 1e-2, 5 )
-#legend = fax.legend( loc='lower left', shadow=True )
-fax.grid()
-plt.savefig( fname, dpi=300 )
-
+print('after final jfactor plot')
